Break base.py into base/ submodule
