+++ properties/base/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+from .base import *
+from .containers import *
+from .instance import *
+from .union import *
